Wrappers around Shortcuts;
The shortcuts need to be available in the app, but I'm keeping a backup copy in this folder of each. We try to load them automatically if they are not found. The unfortunate bit is that these backups may fall out of date with the original thing, but then again they aren't meant to change much